build cleanup
upgrade gradle to 9.7.0
update kotlin to 2.4.10
kotlinx-coroutines = "1.11.0"
xmlutil-version = "1.0.2"

simplify test dependencies

fix minor compiler warnings
fix minor test failures

disable property (kotest) tests for now
document new directory layout

+++ core/src/test/kotlin/org/cryptobiotic/rlauxe/betting/TestGABetting2.kt
@@ -29,7 +29,7 @@ class TestGABetting2 {
             eta = 0.5,
             noerror = noerror,
             riskLimit=alpha,
-            Npop = Nc.toInt(),
+            Npop = Nc,
             Nphantoms=0,
         )
